Use Int32 for string buffer

See merge request common/KM3io.jl!66

+++ src/hardware.jl
@@ -551,7 +551,7 @@ function read_detx(io::IO)
 
     modules = Dict{Int32, DetectorModule}()
     locations = Dict{Tuple{Int, Int}, DetectorModule}()
-    strings = Int8[]
+    strings = Int32[]
     _pmt_id_module_map = Dict{Int, DetectorModule}()
 
     # a counter to work around the floor == -1 bug in some older DETX files
